requirements.txt
asgiref==3.8.1
asttokens==3.0.0
colorama==0.4.6
comm==0.2.2
debugpy==1.8.11
decorator==5.1.1
Django==5.2
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
executing==2.1.0
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
mysqlclient==2.2.7
nest-asyncio==1.6.0
packaging==24.2
parso==0.8.4
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.1
pure_eval==0.2.3
Pygments==2.19.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pywin32==308
pyzmq==26.2.0
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
wcwidth==0.2.13
wheel==0.45.1
